xulrunner 1.9.3: Fix assumption that jsid == jsval
The two have always been conceptually distinct types.  Even in 1.9.3,
they are still the same in implementation, but to avoid a pile of
warnings, we should avoid confusing them. Unfortunately, even the
SpiderMonkey docs are wrong in a few places, so gjs' code confusing
them is understandable.

Anyways, fix up places that made this assumption.