Drop load and call contexts
Apparently for XULRunner embedding reasons, multiple different
contexts were added to gjs.  We're using these in a way not really
intended by Spidermonkey.  The default intended design is that there's
one (active) JSContext * per thread.

In particular, stacking JS_BeginRequest(context_a);
JS_BeginRequest(context_b) is "dubious" in current Spidermonkey.

For the load context: If we want to e.g. avoid having modules loaded
multiple times even if run from a non-default context (and thus a
non-default global object), the obvious implementation technique is to
just hook them off the GjsRuntime default global object.

For the callback context: We need to assume that we have a standard
context provided by Gjs, which is the thread default one.  Actually,
one possibility here is to root the global object for a given
callback.

https://bugzilla.gnome.org/show_bug.cgi?id=622896