Drop load and call contexts
Apparently for XULRunner embedding reasons, multiple different
contexts were added to gjs.  We're using these in a way not really
intended by Spidermonkey.  The default intended design is that there's
one JSContext * per thread.

In particular, stacking contexts is "dubious" in current Spidermonkey.

The stated rationale for the load and call contexts is unconvincing.

For the load context: If we want to e.g. avoid having modules loaded
multiple times even if run from a non-default context (and thus a
non-default global object), the obvious implementation technique is to
just hook them off the GjsRuntime default global object.

For the callback context: The rationale is just busted; we should
always use the default thread context, end of story.