GDBus: allow overriding _init in a Gio.DBusProxyClass
A Gio.DBusProxyClass is a Lang.Class and should behave like one,
including the _init handling. In particular, this allows for
"slim proxies", that can be created without passing parameters
(or with a subset of them).
This is implemented by moving the special initialization of DBusProxy
to an _init override, instead of hooking into init and init_async methods.

https://bugzilla.gnome.org/show_bug.cgi?id=669350