Rework test suite to be installed
The test suite moves to a subdirectory "installed-tests" that
is configured and built indepenently of the main source.

This allows us to automate running it much more easily; the tests run
in qemu and a logged in desktop session.  Thus the test suite doesn't
have to launch its own session bus and Xvfb for example.

It was necessary to change the test suite to do JUnit as a regular
import; the previous method required a custom launcher process, but
it's a lot saner to make the test cases just plain gjs code.

See https://live.gnome.org/GnomeGoals/InstalledTests