coverage: Add GjsCoverage
GjsCoverage is a wrapper around coverage.js, a new module
which uses the JS Debugger API and JS Reflect API to collect
hits counts for lines, branches and functions.

Pass -C or --coverage-path to specify a file that should be
added to the coverage report. --coverage-output must also
be passed when using these options and specifies a directory
to write coverage reports to. This will copy the directory
structure of covered files into this directory and generate
an lcov compatible coverage report at coverage.lcov. This file
will be opened in append-mode and should be deleted when coverage
reports need to be regenerated.