coverage: Enable IonMonkey and BaselineJIT in coverage mode
This should give tests running in coverage mode a slight
speedup.

Previously, we just disabled the JIT during coverage mode, as
it was crashing in the JIT. Upon further analysis, it appears
as though the crash is a bug in js24
(https://bugzilla.mozilla.org/show_bug.cgi?id=1120934). We keep
a single runtime around for a thread's lifetime, which means that
in most cases it is never destroyed. However, trace callbacks
were not correctly added for some internal JIT structures inside
the runtime, which means that if a context was destroyed, the
garbage collector could reclaim that JIT code, but not the
IonRuntime which owned it.

When the next context was created, upon entering a compartment
it would check if an IonRuntime was available, and only
regenerate that JIT code if it was unavailable. If it was
available, then mozjs simply assumed that it was valid and
jumped to it. This caused a jump to an invalid address.

In appears as though the only way to remedy this problem is to
destroy the runtime and re-create it when needed.

gjs_clear_thread_runtime was provided for that purpose. It should
be called whenever a context has been destroyed, but the caller
intends to create a new context later - it will also explictly
clear the runtime.

Fixes #742852